--- cloud/go.mod
@@ -1,15 +1,0 @@
-module github.com/dc-lab/sky/cloud
-
-go 1.13
-
-require (
-	github.com/aws/aws-sdk-go v1.30.16
-	github.com/dc-lab/sky/api/proto/cloud v0.0.0-20200503102138-2d5b83e48e13
-	github.com/dc-lab/sky/api/proto/common v0.0.0-20200503100219-97ef56d66353
-	github.com/go-chi/chi v4.1.1+incompatible
-	github.com/go-chi/render v1.0.1
-	github.com/spf13/cobra v1.0.0
-	github.com/spf13/pflag v1.0.3
-	github.com/swaggo/http-swagger v0.0.0-20200308142732-58ac5e232fba
-	google.golang.org/grpc v1.29.1
-)
