+++ cloud_manager/client/helpers/grpc_call_routine.go
