+++ cloud_manager/client/helpers/response_validation.go
