+++ cloud_manager/client/main.go
