+++ cloud_manager/go.mod
@@ -1,0 +1,12 @@
+module github.com/dc-lab/sky/cloud_manager
+
+go 1.13
+
+require (
+	github.com/aws/aws-sdk-go v1.30.24
+	github.com/dc-lab/sky/api/proto/cloud v0.0.0-20200511135947-d28b28ab3cf3
+	github.com/dc-lab/sky/api/proto/common v0.0.0-20200511142319-b9afa317e44d
+	github.com/spf13/cobra v1.0.0
+	github.com/spf13/pflag v1.0.5
+	google.golang.org/grpc v1.29.1
+)
