+++ cloud_manager/server/handlers/allocate_instance.go
