+++ cloud_manager/server/handlers/connect_instance.go
