+++ cloud_manager/server/handlers/deallocate_instance.go
