+++ cloud_manager/server/handlers/deploy_image.go
