+++ cloud_manager/server/handlers/disconnect_instance.go
