+++ cloud_manager/server/handlers/start_instance.go
