+++ cloud_manager/server/handlers/stop_instance.go
